\name{ABAB}

\alias{ABAB}

\docType{data}

\title{Hypothetical ABAB data}

\description{Hypothetical data from Onghena (1992), collected in an ABAB design with 24 measurement times.}

\usage{data(ABAB)}

\format{
  A data frame with 24 observations.
  \itemize{
    \item The first column contains the condition/phase labels ("A1", "B1", "A2", "B2").
    \item The second column contains the obtained scores.
    \item The rows and columns are not labeled.
  }
}

\source{Onghena, P. (1992). Randomization tests for extensions and variations of AB single-case experimental designs: A rejoinder. Behavioral Assessment, 14, 153-171.}

\references{Bulte, I., & Onghena, P. (2008). An R package for single-case randomization tests. Behavior Research Methods, 40, 467-478.}

\examples{data(ABAB)}

\keyword{datasets}
