\name{SCRT-package}
\alias{SCRT-package}
\docType{package}
\title{
Single-Case Randomization Tests
}
\description{
Design single-case phase, alternation and multiple-baseline experiments, and conduct randomization tests on data gathered by means of such designs.
}
\details{
\tabular{ll}{
Package: \tab SCRT\cr
Type: \tab Package\cr
Version: \tab 1.3.1\cr
Date: \tab 2019-12-20\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Isis Bulte, Tamal Kumar De, Patrick Onghena

Maintainer: Tamal Kumar De <tamalkumar.de@kuleuven.be>
}

\keyword{ package }
