\name{SCRT-package}
\alias{SCRT-package}
\docType{package}
\title{
Single-Case Randomization Tests
}
\description{
A package for designing single-case phase, alternation and multiple-baseline experiments, as well as for conducting randomization tests on data gathered by means of such designs. 
}
\details{
\tabular{ll}{
Package: \tab SCRT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-13\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Isis Bulte and Patrick Onghena

Maintainer: Marlies Vervloet <marlies.vervloet@ppw.kuleuven.be>
}

\keyword{ package }
